rename: color 패키지로 BoardColorAdapter와 BoardColorViewHolder 이동
